pyproject.toml
[project]
name = "ingenieria-web-grupo2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.7",
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "django-allauth>=65.11.2",
    "django-cloudinary-storage>=0.3.0",
    "django-widget-tweaks>=1.5.0",
    "gunicorn>=23.0.0",
    "django-haystack>=3.2.1",
    "whoosh>=2.7.4",
    "oauthlib>=3.3.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "six>=1.17.0",
    "whitenoise[brotli]>=6.9.0",
]
